Add models for D415, D435, D435i.
For visualization, can be presented using view_model.launch.py
