+++ realsense2_description/urdf/_d435.urdf.xacro
@@ -144,7 +144,7 @@ aluminum peripherial evaluation case.
     </xacro:if>
 
     <xacro:if value="$(arg add_plug)">
-      <xacro:usb_plug parent="${name}_link">
+      <xacro:usb_plug parent="${name}_link" name="${name}_usb_plug">
         <origin xyz="${d435_cam_mount_from_center_offset - 0.02095} ${-d435_cam_depth_py - 0.0353} 0" rpy="0 0 0"/>
       </xacro:usb_plug>
     </xacro:if>
